修正了色欲Boss的动画 优化了镜头抖动 添加了装备 暴怒 武器时会不断减少Light值的逻辑(测试)

+++ Assets/Scripts/Commands/PlayerLightChangeCommand.cs
@@ -2,9 +2,9 @@ using QFramework;
 
 public class PlayerLightChangeCommand : AbstractCommand
 {
-    int _lightChange;
+    float _lightChange;
 
-    public PlayerLightChangeCommand(int lightChange)
+    public PlayerLightChangeCommand(float lightChange)
     {
         _lightChange = lightChange;
     }
